@@ -12,43 +12,51 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^5.2.3",
-    "@angular/cdk": "^5.1.1",
-    "@angular/common": "^5.2.0",
-    "@angular/compiler": "^5.2.0",
-    "@angular/core": "^5.2.0",
-    "@angular/forms": "^5.2.0",
-    "@angular/http": "^5.2.0",
-    "@angular/material": "^5.1.1",
-    "@angular/platform-browser": "^5.2.0",
-    "@angular/platform-browser-dynamic": "^5.2.0",
-    "@angular/router": "^5.2.0",
-    "@ng-bootstrap/ng-bootstrap": "^1.0.0",
+    "@angular/animations": "^7.0.2",
+    "@angular/cdk": "^7.0.2",
+    "@angular/common": "^7.0.2",
+    "@angular/compiler": "^7.0.2",
+    "@angular/core": "^7.0.2",
+    "@angular/forms": "^7.0.2",
+    "@angular/http": "^7.0.2",
+    "@angular/material": "^7.0.2",
+    "@angular/platform-browser": "^7.0.2",
+    "@angular/platform-browser-dynamic": "^7.0.2",
+    "@angular/router": "^7.0.2",
+    "@ng-bootstrap/ng-bootstrap": "^4.0.0",
     "angular-material": "^1.1.6",
+    "bcrypt": "^1.0.3",
+    "bcrypt-nodejs": "0.0.3",
     "body-parser": "^1.18.2",
     "bootstrap": "^4.0.0-beta.3",
     "core-js": "^2.4.1",
     "file-saver": "^1.3.3",
+    "jsonwebtoken": "^8.2.0",
     "material-design-icons": "^3.0.1",
-    "mongoose": "^5.0.3",
+    "mongoose": "^5.0.7",
+    "morgan": "^1.9.0",
     "multer": "^1.3.0",
     "ng2-file-upload": "^1.3.0",
     "ngx-bootstrap": "^2.0.2",
-    "ngx-chips": "^1.6.3-rc.3",
-    "npm": "^5.6.0",
-    "rxjs": "^5.5.6",
-    "zone.js": "^0.8.19"
+    "ngx-chips": "^2.0.0-beta.0",
+    "npm": "^5.7.1",
+    "passport": "^0.4.0",
+    "passport-jwt": "^3.0.1",
+    "rxjs": "^6.3.3",
+    "shelljs": "^0.8.1",
+    "zone.js": "^0.8.26"
   },
   "devDependencies": {
-    "@angular/cli": "1.6.5",
-    "@angular/compiler-cli": "^5.2.0",
-    "@angular/language-service": "^5.2.0",
+    "@angular-devkit/build-angular": "~0.10.0",
+    "@angular/cli": "^7.0.4",
+    "@angular/compiler-cli": "^7.0.2",
+    "@angular/language-service": "^7.0.2",
     "@types/file-saver": "^1.3.0",
     "@types/jasmine": "~2.8.3",
     "@types/jasminewd2": "~2.0.2",
     "@types/node": "~6.0.60",
     "angular-ide": "^0.9.39",
-    "codelyzer": "^4.0.1",
+    "codelyzer": "^4.5.0",
     "jasmine-core": "~2.8.0",
     "jasmine-spec-reporter": "~4.2.1",
     "karma": "~2.0.0",
@@ -59,6 +67,6 @@
     "protractor": "~5.1.2",
     "ts-node": "~4.1.0",
     "tslint": "~5.9.1",
-    "typescript": "~2.5.3"
+    "typescript": "~3.1.6"
   }
 }
