@@ -1,5 +1,6 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
+import { AuthGuardService } from './services/auth-guard.service';
 
 import { PhotoComponent } from './photo/photo.component';
 import { AcceuilComponent } from './acceuil/acceuil.component';
@@ -7,7 +8,7 @@ import { AcceuilComponent } from './acceuil/acceuil.component';
 const routes: Routes = [
   { path: '', redirectTo: '/acceuil', pathMatch: 'full' },
   { path: 'acceuil', component: AcceuilComponent },
-  { path: 'photo/:album', component: PhotoComponent }
+  { path: 'photo/:album', component: PhotoComponent, canActivate: [AuthGuardService] }
 ];
 
 @NgModule({
